Added the glue for incorporating generic model TMs into the framework: - allow specification of pre-trained TM files in Makefile.params using absolute or relative pathnames. - automatically create the necessary symbolic links in the framework. - pre-trained TMs can be used for either regular tms or mixtms. - error checking that the linked files exist.
Allow the user to explicitly specify the training corpus with word alignment models to use for training the jpt to train the mixtm mixture weights.

Allow the (advanced) user to explicitly specify the dev set to use for tuning the mixtm mixture weights (default: ${TUNE_DECODE}).

Fixed the generic model LM glue to do better error checking, and to be
consistent with the way it is done for TMs.

Fixed some comments in Makefile.params.