Add TMS_WITH_ALIGNMENT_FIELD option and document USE_SIGPRUNING.
1) TMS_WITH_ALIGNMENT_FIELD turns on keeping the alingment information in
phrase tables, using the a= field.  This works with all types of phrase tables,
with or without sig pruning, in TextPTs as well as TPPTs.

The TMS_WITH_ALIGNMENT_FIELD option is in the top-level Makefile.params, in the
advanced parameter section.  We're going to depend on the a= field for
truecasing soon, and it is used by various features.  It does not cost that
much to generate or carry around, so why not have it there by default.

2) Document the USE_SIGPRUNING option in the top-level Makefile.params, in the
user variable section.