PTGSH-274: Fixed BiLM training for IBM4 word alignment.
The problem is that IBM4 training filters the corpus first; the filtered
corpus needs to be used for the bitokens files for the BiLM training,
not the original corpus; this is also the case for the TUNE_BILM
bitokens file.

Also, fixed a bug that omitted the tuning of the LM discount parameters
for BiLMs because the dependency of the LM file on the TUNE_BILM
bitokens file was not specified correctly. This also required fixing the
workflow for training the TUNE_BILM bitokens files.

Various clean up in the makefile, including making all derived corpus
files for the TUNE_BILM be uncompressed, while those for TRAIN_BILM are
compressed. Also arranged that word2class and word_align_tool are called
locally for uncompressed (TUNE_BILM) files but psub'ed (through
run-parallel.sh) for compressed (TRAIN_BILM) files.